refactor: 테스트클래스 패키지 위치 수정
